Refactored meta structure to base, models, train structure
