+++ src/orc/classifier/__init__.py
@@ -1,0 +1,6 @@
+"""Classification with Reservoir Computers.
+
+This module is currently a placeholder for future classifier implementations.
+"""
+
+__all__ = []
