+++ src/orc/classifier/base.py
@@ -1,0 +1,6 @@
+"""Base classes for Reservoir Computer Classifiers.
+
+This module is currently a placeholder for future classifier implementations.
+"""
+
+# TODO: Implement RCClassifierBase and related classes
