+++ src/orc/classifier/models.py
@@ -1,0 +1,6 @@
+"""Classifier models based on Reservoir Computing.
+
+This module is currently a placeholder for future classifier implementations.
+"""
+
+# TODO: Implement classifier models (e.g., ESNClassifier)
