+++ src/orc/classifier/train.py
@@ -1,0 +1,6 @@
+"""Training functions for Reservoir Computer Classifiers.
+
+This module is currently a placeholder for future classifier implementations.
+"""
+
+# TODO: Implement training functions for classifiers
