+++ src/orc/control/__init__.py
@@ -1,0 +1,6 @@
+"""Control with Reservoir Computers.
+
+This module is currently a placeholder for future control implementations.
+"""
+
+__all__ = []
