+++ src/orc/control/base.py
@@ -1,0 +1,6 @@
+"""Base classes for Reservoir Computer Controllers.
+
+This module is currently a placeholder for future control implementations.
+"""
+
+# TODO: Implement RCControllerBase and related classes
