+++ src/orc/control/models.py
@@ -1,0 +1,6 @@
+"""Controller models based on Reservoir Computing.
+
+This module is currently a placeholder for future control implementations.
+"""
+
+# TODO: Implement controller models (e.g., ESNController)
