+++ src/orc/control/train.py
@@ -1,0 +1,6 @@
+"""Training functions for Reservoir Computer Controllers.
+
+This module is currently a placeholder for future control implementations.
+"""
+
+# TODO: Implement training functions for controllers
