--- src/orc/models/__init__.py
@@ -1,21 +1,0 @@
-"""Implementations of common model architectures."""
-
-from orc.models import esn
-from orc.models.esn import (
-    CESNForecaster,
-    EnsembleESNForecaster,
-    ESNForecaster,
-    train_CESNForecaster,
-    train_EnsembleESNForecaster,
-    train_ESNForecaster,
-)
-
-__all__ = [
-    "esn",
-    "ESNForecaster",
-    "CESNForecaster",
-    "train_ESNForecaster",
-    "train_CESNForecaster",
-    "train_EnsembleESNForecaster",
-    "EnsembleESNForecaster",
-]
