+++ src/orc/tuning/__init__.py
@@ -1,0 +1,6 @@
+"""Hyperparameter tuning utilities for Reservoir Computers.
+
+This module is currently a placeholder for future tuning implementations.
+"""
+
+__all__ = []
